Extract methods to separate file
