+++ Pylette/src/color_extraction.py
@@ -8,11 +8,10 @@ import numpy as np
 import requests  # type: ignore
 from numpy.typing import NDArray
 from PIL import Image
-from sklearn.cluster import KMeans
 
-from Pylette.src.color import Color
+from Pylette.src.extractors.k_means import k_means_extraction
+from Pylette.src.extractors.median_cut import median_cut_extraction
 from Pylette.src.palette import Palette
-from Pylette.src.utils import ColorBox
 
 ImageType_T: TypeAlias = Union["os.PathLike[Any]", bytes, NDArray[float], str]
 
@@ -23,35 +22,6 @@ class ImageType(str, Enum):
     ARRAY = "array"
     URL = "url"
     NONE = "none"
-
-
-def median_cut_extraction(arr: np.ndarray, height: int, width: int, palette_size: int) -> list[Color]:
-    """
-    Extracts a color palette using the median cut algorithm.
-
-    Parameters:
-        arr (np.ndarray): The input array.
-        height (int): The height of the image.
-        width (int): The width of the image.
-        palette_size (int): The number of colors to extract from the image.
-
-    Returns:
-        list[Color]: A list of colors extracted from the image.
-    """
-
-    arr = arr.reshape((width * height, -1))
-    c = [ColorBox(arr)]
-
-    # Each iteration, find the largest box, split it, remove original box from list of boxes, and add the two new boxes.
-    while len(c) < palette_size:
-        largest_c_idx = np.argmax(c)
-        # add the two new boxes to the list, while removing the split box.
-        c = c[:largest_c_idx] + c[largest_c_idx].split() + c[largest_c_idx + 1 :]
-
-    total_pixels = width * height
-    colors = [Color(tuple(map(int, box.average)), box.pixel_count / total_pixels) for box in c]
-
-    return colors
 
 
 def _parse_image_type(image: ImageType_T) -> ImageType:
@@ -169,28 +139,3 @@ def request_image(image_url: str) -> Image.Image:
         return img
     else:
         raise ValueError("The URL did not point to a valid image.")
-
-
-def k_means_extraction(arr: NDArray[float], height: int, width: int, palette_size: int) -> list[Color]:
-    """
-    Extracts a color palette using KMeans.
-
-    Parameters:
-        arr (NDArray[float]): The input array.
-        height (int): The height of the image.
-        width (int): The width of the image.
-        palette_size (int): The number of colors to extract from the image.
-
-    Returns:
-        list[Color]: A palette of colors sorted by frequency.
-    """
-    arr = np.reshape(arr, (width * height, -1))
-    model = KMeans(n_clusters=palette_size, n_init="auto", init="k-means++", random_state=2024)
-    labels = model.fit_predict(arr)
-    palette = np.array(model.cluster_centers_, dtype=int)
-    color_count = np.bincount(labels)
-    color_frequency = color_count / float(np.sum(color_count))
-    colors = []
-    for color, freq in zip(palette, color_frequency):
-        colors.append(Color(color, freq))
-    return colors
